release 1.0
Source code + application (jar file). Final release (1.0).
Features : indexation, updating/deleting of an index, help to
configuration.
