+++ agrold_indexation/application/config.properties
@@ -1,0 +1,9 @@
+#Tue Jul 26 15:44:54 CEST 2016
+shield.ssl.keystore.path=
+server.port=9300
+shield.ssl.truststore.password=
+shield.ssl.truststore.path=
+cluster.name=AgroLD
+shield.ssl.keystore.password=
+server.host=localhost
+shield.user=test_user:changeme
