--- agrold_indexation/pom.xml
@@ -1,54 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>ibc.agrold</groupId>
-  <artifactId>agrold.indexation</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <name>agrold.indexation</name>
-  <url>http://maven.apache.org</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-  	   <dependency>
-    		<groupId>junit</groupId>
-    		<artifactId>junit</artifactId>
-    		<version>4.12</version>
-	   </dependency>
-	   <dependency>
-	   		<groupId>org.elasticsearch</groupId>
-	   		<artifactId>elasticsearch</artifactId>
-	    	<version>2.2.1</version>
-	   </dependency>
-	   <dependency>
-		    <groupId>org.apache.jena</groupId>
-		    <artifactId>jena-core</artifactId>
-		    <version>3.1.0</version>
-		</dependency>
-			<dependency>
-        	<groupId>org.json</groupId>
-        	<artifactId>json</artifactId>
-        	<version>20160212</version>
-        </dependency>
-        <dependency>
-    		<groupId>com.google.code.gson</groupId>
-    		<artifactId>gson</artifactId>
-    		<version>2.6.2</version>
-		</dependency>
-        <dependency>
-    		<groupId>org.slf4j</groupId>
-		    <artifactId>slf4j-nop</artifactId>
-		    <version>1.7.21</version>
-		</dependency>
-		<dependency>
-		    <groupId>org.apache.clerezza.ext</groupId>
-		    <artifactId>org.apache.jena.jena-core</artifactId>
-		    <version>2.11.1_1</version>
-		</dependency>
-   </dependencies>
-</project>
