--- agrold_indexation/src/main/java/indexation/properties/config.properties
@@ -1,9 +1,0 @@
-################### Configuration file (indexation) ###################
-cluster.name=agrold
-server.host=richelais.cirad.fr
-server.port=9300
-shield.user=
-shield.ssl.keystore.path=
-shield.ssl.keystore.password=
-shield.ssl.truststore.path=
-shield.ssl.truststore.password=
